Extract appearable component
